Update 1.11
Added window title into the process list.
Added a check to see if the user has the required tessdata files for the selected config.
Tesseract does no longer get initialized each awake iteration to eat performance.
Logging has been improved to further identify issues with accuracy for the user.
Changed the awake comparison and the awake is no longer required to be an exact match. Now you can simply add a part of the awake in-game text into the config to identify it.
	Example: The in-game awake "Additional Damage of Critical Hits", if "Additional" is written in the config that works.