Updated to version 1.31
Updated tesseract version to 4. It now uses the standalone tesseract 4 executable instead of .NET wrapper.
Added the latest trained tesseract data.
Hotfix for accuracy issues by fixing DPI on the bitmaps.
Fixed the log not printing new lines correctly.
When adding a new awake, the inputs are cleared for the next awake.
Major code cleanup and refactor.
Added a prompt when dragging the awake rectangle that asks if you want to save, discard or cancel when pressing another key than enter.
Added the ability to temporarily hide the awake rectangle selection window to be able to hover over the item in the inventory to properly fit the rectangle by holding TAB.
The rectangle awake selection window is now more aesthetically pleasing.
Added the ability to benchmark and test the accuracy on pre-taken images of awakes for debugging and development purposes. Not available for normal user.
Added better and more error handling. If the awake parser now realizes something is wrong, it stops with an error instead of blindly continuing as it did before.
Added a info messagebox when the user adds one of more of the same awake to the preferred awake list informing the user that there is no need because they are combined internally anyway.
Added the ability to have multiple awakening groups to support the fact if the user wants any of 2 different awakes. In other words, e.g. STR+30 (Group 1) or INT+24 (Group 2), the bot will stop at any of the groups.
Added tooltips on the next awake group feature to futher inform the user how it works.
Added the ability to add ignored words in a server config for the awake parser to ignore. Default added value in the config is "Blessing Options" to awakening of CS set parts. Previously you had to draw the awake rectangle just below the "Blessing Options" text for it to work.
When pressing "END" or stopping the bot, the window will be disabled until it has finished its routine.
Added an error when launching the bot on a computer that has increased or decreased the DPI scaling. That completely makes the bot unfunctionalble because it changes all font qualities.
Tested with 126 different awakes with 100% sucess rate.
Removed the so called "panic thread" and replaced it with a better solution.
Removed the feature that added the ability to only add a part of the awake text in the config to identify it. It caused inconsistencies with other awakes containing the same characters.