Updated to version 1.34
Fixed issue with some servers have bugs that prevent an item description from appearing when simply just moving the mouse over the item quickly. It now move two times after having hovered on the item.
Fixed a bug caused when the OCR engine did not recognize a space between the awake and the value.
Added an option to make the bot stop when the OCR engine does not recognize an awake. Normally it just continues because it assumes its just missing that awake in the server config.