Updated to 1.52
Updated benchmark images to be up to date with new way to parse images.
New way to enlarge an image based on the original image DPI, hopefully tesseract will like that more.
Added the OCR ignore words into the tesseract character whitelist to fix inaccuracy.
If user is missing win32u.dll the bot now falls back to the default winapi calls (note: does not bypass gameguard if that occurs).
Added text_to_find in the config to work as a substring of the awake text because the old way had an issue because it did not whitelist all characters for tesseract for the whole awake.
If the bot now realizes the awake stays the same for 3 tries, it assumes the awakes/reversions are gone and stops.
Added custom dictionary support because Tesseract at its core seems to be for documents, hence its dictionary usage.
Upon bot start, it adds all awakes from the config into its dictionary for more accuracy.
Added a Moon Flyff config to showcase the new text_to_find feature in the config.